added images into code and created and coded the about me page
I also added classes to my div tags so that I can understand what is inside of each div tag and make the website easier to read as a developer. Obviously classes are used mainly for CSS purposes so we don't need it now but this will help us later.